vulkan: separate handle concept from memory-backed handle
When we introduce exportable semaphores, we'll have handles which are
not memory-backed. To avoid a sloppy API, let's explicitly separate
the two ideas. This also presents an opportunity to pull the object
offset within the memory into the mem_handle definition.

This is obviously an API break for any external usage of buffers
or textures.