vulkan: Make pl_gpu_handle a single handle
In practice, it is not going to be useful to export a given resource
as multiple types of external handle - it will just one or another
type.

Accordingly, let's enforce that the pl_gpu_handle is just one value
and that the pl_handle_types specified for a given resource has only
one bit value set.

This is mostly a simple substitution, but we must also note that
vk_slabs must also carry the handle type explicitly, as we can no
longer look at which field(s) are set in a pl_gpu_handle.

This is also an API break for any code that tries to export buffers
or textures.